requirements.txt
attrs==22.1.0
bcrypt==4.0.0
cffi==1.15.1
colorama==0.4.5
commonmark==0.9.1
cryptography==38.0.1
future==0.18.2
graphviz==0.20.1
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
netmiko==4.1.2
ntc-templates==3.1.0
packaging==21.3
paramiko==2.11.0
pexpect==4.8.0
pluggy==1.0.0
pprintpp==0.4.0
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
pytest==7.1.3
pytest-clarity==1.0.1
PyYAML==6.0
rich==12.5.1
scp==0.14.4
six==1.16.0
tabulate==0.8.10
tenacity==8.1.0
textfsm==1.1.2
tomli==2.0.1
typing-extensions==4.3.0
